EVALSYS-1389 Evaluations Unit Tests don't compile with Sakai 10
git-svn-id: https://source.sakaiproject.org/contrib/evaluation/trunk@86560 fdecad78-55fc-0310-b1b2-d7d25cf747c9

+++ impl/src/java/org/sakaiproject/evaluation/dao/EvaluationDaoImpl.java
@@ -1429,7 +1429,7 @@ public class EvaluationDaoImpl extends HibernateGeneralGenericDao implements Eva
                 DetachedCriteria dc = DetachedCriteria.forClass(EvalItem.class).add(
                         Restrictions.eq("locked", Boolean.TRUE)).add(Restrictions.eq("scale.id", scale.getId()))
                         .setProjection(Projections.rowCount());
-                if (((Integer) getHibernateTemplate().findByCriteria(dc).get(0)).intValue() > 0) {
+                if (((Long) getHibernateTemplate().findByCriteria(dc).get(0)).intValue() > 0) {
                     // this is locked by something, we cannot unlock it
                     log.info("Cannot unlock scale (" + scale.getId() + "), it is locked elsewhere");
                     return false;
